LayoutTests/inspector/debugger/break-on-exception-window-onerror-expected.txt

CONSOLE MESSAGE: line 7: throwing TestError
CONSOLE MESSAGE: line 8: window.onerror: TestError
CONSOLE MESSAGE: line 9: DONE
CONSOLE MESSAGE: line 8: TestError
Checking pause locations when pausing on all exceptions.

PAUSE #1 AT: doThrow:8:22
